useDataSummary in useLeaflet + move map utils to composables/utils
